Promote calibration readers and trim PETR dependencies
