+++ examples/onnx_ptq/petr/mmdet3d-nuscenes-only.patch
@@ -1,0 +1,26 @@
+diff --git a/mmdet3d/core/evaluation/__init__.py b/mmdet3d/core/evaluation/__init__.py
+index b1d489f..939e90b 100644
+--- a/mmdet3d/core/evaluation/__init__.py
++++ b/mmdet3d/core/evaluation/__init__.py
+@@ -5 +4,0 @@ from .kitti_utils import kitti_eval, kitti_eval_coco_style
+-from .lyft_eval import lyft_eval
+@@ -9 +8 @@ __all__ = [
+-    'kitti_eval_coco_style', 'kitti_eval', 'indoor_eval', 'lyft_eval',
++    'kitti_eval_coco_style', 'kitti_eval', 'indoor_eval',
+diff --git a/mmdet3d/datasets/__init__.py b/mmdet3d/datasets/__init__.py
+index dab7502..ae58b25 100644
+--- a/mmdet3d/datasets/__init__.py
++++ b/mmdet3d/datasets/__init__.py
+@@ -8 +7,0 @@ from .kitti_mono_dataset import KittiMonoDataset
+-from .lyft_dataset import LyftDataset
+@@ -34 +33 @@ __all__ = [
+-    'build_dataset', 'NuScenesDataset', 'NuScenesMonoDataset', 'LyftDataset',
++    'build_dataset', 'NuScenesDataset', 'NuScenesMonoDataset',
+diff --git a/requirements/runtime.txt b/requirements/runtime.txt
+index 643cb0c..d0c7719 100644
+--- a/requirements/runtime.txt
++++ b/requirements/runtime.txt
+@@ -1 +0,0 @@
+-lyft_dataset_sdk
+@@ -6 +4,0 @@ nuscenes-devkit
+-plyfile
