--- examples/onnx_ptq/petr/requirements-mmdet3d.txt
@@ -1,1 +1,0 @@
-mmdet3d==1.0.0rc6
